Use MUnit in ServerPriorKnowledgeHandshakerSuite
